refactor: admin 모듈의 enums 패키지 경로 수정
