+++ admin/src/main/java/com/kernel/admin/common/enums/BoardType.java
@@ -1,4 +1,4 @@
-package com.kernel.admin.domain.enumerate;
+package com.kernel.admin.common.enums;
 
 public enum BoardType {
     NOTICE("공지"),
